configurations are now classes
